+++ packages/flutter_reload/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.7 - 2024-09-07
 
 - Remove reload lifecycle for GuardViewModelMixin to increase compatibility with other state management.
+- Expose the list count for the pagination model.
 
 ## 0.0.6 - 2024-07-17